mapArrayToButtons aangepast: op regel 27-28 KEY vervangen door NAME omdat button niet werkte. Maar hetzelfde probleem blijft bestaan, laadt pas bij refreshen. Dus mogelijk ligt de oorzaak ergens anders.
