+++ src/helpers/mapArrayToButtons.jsx
@@ -1,14 +1,34 @@
-import {Link} from "react-router-dom";
+// import {Link} from "react-router-dom";
+// import Button from "../components/button/Button.jsx";
+//
+// function mapArrayToButtons(inputArray, typeOfLink) {
+//     console.log(inputArray);
+//     if (!inputArray || inputArray.length === 0) {
+//         return <p>No items found</p>;
+//     }
+//     return inputArray.map(item => (
+//         <Link key={item.key} to={`/${typeOfLink}/${item.key}`}>
+//             <Button>{item.value}</Button>
+//         </Link>
+//     ));
+// }
+// export default mapArrayToButtons;
+import { Link } from "react-router-dom";
 import Button from "../components/button/Button.jsx";
 
 function mapArrayToButtons(inputArray, typeOfLink) {
-    if (!inputArray || inputArray.length === 0) {
+    // Convert the object to an array of [key, value] pairs
+    const arrayFromObject = Object.entries(inputArray);
+
+    if (!arrayFromObject || arrayFromObject.length === 0) {
         return <p>No items found</p>;
     }
-    return inputArray.map(item => (
-        <Link key={item} to={`/${typeOfLink}/${item}`}>
-            <Button>{item}</Button>
+//TODO key is vervangen door name (26-2-2025) omdat button niet werkte. Maar het lijkt een ander probleem te zijnwant moet nog steeds refreshen voordat hij wel goed laadt.
+    return arrayFromObject.map(([name, value]) => (
+        <Link key={name} to={`/${typeOfLink}/${name}`}>
+            <Button>{value}</Button>
         </Link>
     ));
 }
+
 export default mapArrayToButtons;
